feat(DB): Moved routes and util functions to new files
